requirements.txt
blinker==1.4
certifi==2018.8.13
chardet==3.0.4
closure==20121212
coverage==3.7.1
decorator==3.3.2
fakeredis==0.4.0
Flask==0.10
Flask-Assets==0.12
Flask-Cors==2.0.1
Flask-DebugToolbar==0.7.1
Flask-SeaSurf==0.1.13
gevent==1.1.2
greenlet==0.4.14
gunicorn==19.4.1
hiredis==0.1.4
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsmin==2.0.2.post1
Markdown==2.3.1
MarkupSafe==1.0
mock==1.0.1
python-coveralls==2.4.3
python-dateutil==1.5
python-dotenv==0.9.1
PyYAML==3.10
redis==2.9.0
repoze.lru==0.7
requests==2.19.1
Routes==2.4.1
sh==1.12.14
six==1.11.0
Sixpack==2.7.0
statsd==3.2.1
urllib3==1.23
webassets==0.12.1
Werkzeug==0.9.1
yuicompressor==2.4.7
